move chain events code to a spearate file
