+++ processor/abci.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/binary"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"time"
 
@@ -396,107 +395,6 @@ func (app *App) DeliverChainEvent(ctx context.Context, tx abci.Tx) error {
 	}
 
 	return app.processChainEvent(ctx, ce, tx.PubKey())
-}
-
-func (app *App) processChainEvent(ctx context.Context, ce *types.ChainEvent, pubkey []byte) error {
-	// first verify the event was emited by a validator
-	if !app.top.Exists(pubkey) {
-		return ErrChainEventFromNonValidator
-	}
-
-	// ack the new event then
-	if !app.evtfwd.Ack(ce) {
-		// there was an error, or this was already acked
-		// but that's not a big issue we just going to ignore that.
-		return nil
-	}
-
-	// OK the event was newly acknowledged, so now we need to
-	// figure out what to do with it.
-	switch ce.Event.(type) {
-	case *types.ChainEvent_Builtin:
-		return app.processChainEventBuiltinAsset(ctx, ce)
-	case *types.ChainEvent_Erc20:
-		return app.processChainEventERC20(ctx, ce)
-	case *types.ChainEvent_Btc:
-		return errors.New("BTC Event not implemented")
-	case *types.ChainEvent_Validator:
-		return errors.New("Validator Event not implemented")
-	default:
-		return ErrUnsupportedChainEvent
-	}
-}
-
-func (app *App) processChainEventBuiltinAsset(ctx context.Context, ce *types.ChainEvent) error {
-	evt := ce.GetBuiltin()
-	if evt == nil {
-		return ErrNotABuiltinAssetEvent
-	}
-
-	switch act := evt.Action.(type) {
-	case *types.BuiltinAssetEvent_Deposit:
-		if err := app.checkVegaAssetID(act.Deposit, "BuiltinAsset.Deposit"); err != nil {
-			return err
-		}
-		return app.banking.DepositBuiltinAsset(act.Deposit, ce.Nonce)
-	case *types.BuiltinAssetEvent_Withdrawal:
-		if err := app.checkVegaAssetID(act.Withdrawal, "BuiltinAsset.Withdrawal"); err != nil {
-			return err
-		}
-		return app.col.Withdraw(ctx, act.Withdrawal.PartyID, act.Withdrawal.VegaAssetID, act.Withdrawal.Amount)
-	default:
-		return ErrUnsupportedEventAction
-	}
-}
-
-func (app *App) processChainEventERC20(ctx context.Context, ce *types.ChainEvent) error {
-	evt := ce.GetErc20()
-	if evt == nil {
-		return ErrNotAnERC20Event
-	}
-
-	switch act := evt.Action.(type) {
-	case *types.ERC20Event_AssetList:
-		if err := app.checkVegaAssetID(act.AssetList, "ERC20.AssetList"); err != nil {
-			return err
-		}
-		// now check that the notary is GO for this asset
-		_, ok := app.notary.IsSigned(
-			act.AssetList.VegaAssetID,
-			types.NodeSignatureKind_NODE_SIGNATURE_KIND_ASSET_NEW)
-		if !ok {
-			return ErrChainEventAssetListERC20WithoutEnoughSignature
-		}
-		return app.banking.EnableERC20(ctx, act.AssetList, evt.Block, evt.Index)
-	case *types.ERC20Event_AssetDelist:
-		return errors.New("ERC20.AssetDelist not implemented")
-	case *types.ERC20Event_Deposit:
-		if err := app.checkVegaAssetID(act.Deposit, "ERC20.AssetList"); err != nil {
-			return err
-		}
-		return app.banking.DepositERC20(act.Deposit, evt.Block, evt.Index)
-	case *types.ERC20Event_Withdrawal:
-		return errors.New("ERC20.Withdrawal not implemented")
-	default:
-		return ErrUnsupportedEventAction
-	}
-}
-
-type HasVegaAssetID interface {
-	GetVegaAssetID() string
-}
-
-func (app *App) checkVegaAssetID(a HasVegaAssetID, action string) error {
-	id := a.GetVegaAssetID()
-	_, err := app.assets.Get(id)
-	if err != nil {
-		app.log.Error("invalid vega asset ID",
-			logging.String("action", action),
-			logging.Error(err),
-			logging.String("asset-id", id))
-		return err
-	}
-	return nil
 }
 
 func (app *App) onTick(ctx context.Context, t time.Time) {
